#27 allow more parameters in constructors
Increased the number of parameters allowed in constructors to 15. This
is because some response DTOs have many fields and no optional
parameters, and are intended to be immutable. It may be worth
considering a less flat structure in DTOs in future that might reduce
the number of arguments required. A builder pattern is probably not
appropriate because none of the parameters are optional.