#30 represent multiple OIDC claims
This class in conjunction with OidcIdentifierClaim allows multiple claim
name-value pairs to be represented as part of an OIDC identifier.
Because OidcIdentifierClaim is an @ElementCollection this class has been
converted to an @Entity (JPA requires this) but will still be treated as an
aggregate value object rather than a conceptual entity.
Removed implementation of Comparable. Fully meeting the contract of
equals and comparable when there is an embedded set is complicated and
it's not clear what the natural order beyond issuer should be. Sorting
is only really required for presentation so these objects can be sorted
by the response DTO factory if required using a Comparator.