#30 revised Subscriber--OidcIdentifier association
This needs to be a bi-directional association so subscriber field added
to OidcIdentifier and included in constructor arguments. Note that this
field is not accessible and has no bearing on object equality because
objects of this class are intended for use as value objects associated
with a particular Subscriber aggregate. This field is only included
because of the requirements of JPA.
Updated tests.